Enemies And Obstalces
The enemies now have clear line of sight, so the player could sneak around them and such. The player will also be pinned  for a couple seconds if the enemy reaches them. The obstacles currently have either normal or trip (as we don't have checkpoints for the trap) which will slow down the player, the normal breaking if the player is going fast enough and the trip always slowing the player, more so if they are going fast.